feat: enhance CV data structure and improve component rendering
Added technical stack information to CV experiences and implemented a function to dynamically generate skills based on experience data. Updated the CV components for better layout and styling, including adjustments to the Address, Education, and ExperienceItem components for improved readability and consistency. Also, included the email in the Nuxt configuration for dynamic rendering across components.